feat: Support rating current playing song
Using the rating buttons in the player controls,
users can send ratings for the current song.

The feature is implemented such that future adapters that do not
support ratings won't show the rating widget. Also if the adapter
supports showing but not setting then clicking on the ratings is
disabled.