Fix Playlist song list updating
When updating the song list in a playlist, we get a list of additions and
removals (in the form of table indexes) to do. We then need to do the removals
first to avoid invalidating the table indexes. The previous code did the
additions first, causing exceptions and bogus songs left in the UI.